refactor: Separate Board in partial classes
